+++ src/ts/index.ts
@@ -37,7 +37,7 @@ import type {
 import type { EffectorInterface } from "./types/services/effector.type";
 import { Libs } from "./utils/libs";
 
-const iVersion = "1.1.4" as const;
+const iVersion = "1.1.5" as const;
 const iName = "SelectiveUI" as const;
 
 const SECLASS = new Selective();